moving Errors and OAuth2Manager to lib/models folder
